Compute mean, confidence interval (?), and covariate balance in Full Gaussian Process (GP)

Calculates the induced covariate balance associated with one hyper-parameter configuration in full GP.

Usage

compute_m_sigma(hyperparam, data, w, GPS_m, kernel_fn = function(x) exp(-x^2))

Arguments

hyperparam

A vector of values of hyper-parameters.

First element: alpha

Second element: beta

Third element: g_sigma (gamma/sigma)

data

A data.table containing all data including outcome, exposure and covariates. In the following order:

Column 1: Outcome (Y)

Column 2: Exposure or treatment (w)

Column 3~m: Confounders (C)

w

A vector of exposure levels at which the CERF is estimated.

GPS_m

A data.table of GPS vectors.

Column 1: A vector of estimated GPS evaluated at the observed exposure levels.

Column 2: Estimated conditional means of the exposure given covariates for all samples (e_gps_pred).

Column 3: Estimated conditional standard deviation of the exposure given covariates for all samples (e_gps_std).

kernel_fn

The covariance function of GP.

Value

A list containing two elements: 1) a vector of absolute weighted correlation of each covariate to the exposure, which is the metric for covariate balance and 2) the estimated CERF at w.all based on the hyper-parameter values in param.

Examples

set.seed(912)
data <- generate_synthetic_data(sample_size = 250, gps_spec = 3)

w.all = seq(0,20,0.1)

data.table::setDT(data)

#Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:8.427888 
#> [2]	train-rmse:6.576370 
#> [3]	train-rmse:5.230828 
#> [4]	train-rmse:4.278385 
#> [5]	train-rmse:3.617113 
#> [6]	train-rmse:3.117391 
#> [7]	train-rmse:2.677701 
#> [8]	train-rmse:2.408475 
#> [9]	train-rmse:2.268593 
#> [10]	train-rmse:2.135648 
#> [11]	train-rmse:1.998507 
#> [12]	train-rmse:1.922534 
#> [13]	train-rmse:1.871516 
#> [14]	train-rmse:1.679066 
#> [15]	train-rmse:1.484187 
#> [16]	train-rmse:1.383167 
#> [17]	train-rmse:1.327371 
#> [18]	train-rmse:1.265584 
#> [19]	train-rmse:1.182439 
#> [20]	train-rmse:1.116496 
#> [21]	train-rmse:1.098769 
#> [22]	train-rmse:1.075959 
#> [23]	train-rmse:0.971201 
#> [24]	train-rmse:0.909766 
#> [25]	train-rmse:0.853721 
#> [26]	train-rmse:0.785449 
#> [27]	train-rmse:0.720735 
#> [28]	train-rmse:0.671513 
#> [29]	train-rmse:0.634115 
#> [30]	train-rmse:0.619789 
#> [31]	train-rmse:0.587799 
#> [32]	train-rmse:0.555885 
#> [33]	train-rmse:0.509312 
#> [34]	train-rmse:0.492669 
#> [35]	train-rmse:0.476967 
#> [36]	train-rmse:0.458201 
#> [37]	train-rmse:0.422026 
#> [38]	train-rmse:0.397187 
#> [39]	train-rmse:0.372963 
#> [40]	train-rmse:0.340009 
#> [41]	train-rmse:0.318249 
#> [42]	train-rmse:0.298455 
#> [43]	train-rmse:0.283954 
#> [44]	train-rmse:0.263343 
#> [45]	train-rmse:0.237295 
#> [46]	train-rmse:0.211401 
#> [47]	train-rmse:0.197621 
#> [48]	train-rmse:0.187469 
#> [49]	train-rmse:0.175141 
#> [50]	train-rmse:0.169089 

tune_res <- compute_m_sigma(hyperparam = c(0.09, 0.09, 10),
                            data = data,
                            w = w.all,
                            GPS_m = GPS_m)

gp.cerf <- tune_res$est