Calculate Posterior Standard Deviations for nnGP Model

Calculates the posterior standard deviation of a point on the CERF based on the nnGP model.

Usage

compute_posterior_sd_nn(
  hyperparam,
  w,
  GPS_w,
  obs_ord,
  sigma2,
  n_neighbor = 10,
  expand = 1
)

Arguments

hyperparam

Values of hyperparameters in the GP model.

w

The exposure level for the point of interest on the CERF.

GPS_w

The GPS for all samples when their exposure levels are set at w.

obs_ord

A matrix of two columns. First column is the observed exposure levels of all samples; second is the GPS at the observed exposure levels. The rows are in ascending order for the first column.

sigma2

A scaler representing sigma^2

n_neighbor

Number of nearest neighbours on one side (see also expand).

expand

Scaling factor to determine the total number of nearest neighbours. The total is 2*expand*n_neighbor.

Value

The posterior standard deviation of the estimated CERF at w.

Examples

set.seed(3089)
data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)

# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:7.977744 
#> [2]	train-rmse:6.473513 
#> [3]	train-rmse:5.278797 
#> [4]	train-rmse:4.420680 
#> [5]	train-rmse:3.853461 
#> [6]	train-rmse:3.509004 
#> [7]	train-rmse:3.157153 
#> [8]	train-rmse:2.889879 
#> [9]	train-rmse:2.738637 
#> [10]	train-rmse:2.645362 
#> [11]	train-rmse:2.533985 
#> [12]	train-rmse:2.409443 
#> [13]	train-rmse:2.083127 
#> [14]	train-rmse:1.919730 
#> [15]	train-rmse:1.793337 
#> [16]	train-rmse:1.656956 
#> [17]	train-rmse:1.432716 
#> [18]	train-rmse:1.390796 
#> [19]	train-rmse:1.330662 
#> [20]	train-rmse:1.205500 
#> [21]	train-rmse:1.067686 
#> [22]	train-rmse:0.967307 
#> [23]	train-rmse:0.860374 
#> [24]	train-rmse:0.783083 
#> [25]	train-rmse:0.744678 
#> [26]	train-rmse:0.691237 
#> [27]	train-rmse:0.636118 
#> [28]	train-rmse:0.605575 
#> [29]	train-rmse:0.563186 
#> [30]	train-rmse:0.543300 
#> [31]	train-rmse:0.527330 
#> [32]	train-rmse:0.473954 
#> [33]	train-rmse:0.437780 
#> [34]	train-rmse:0.403503 
#> [35]	train-rmse:0.358255 
#> [36]	train-rmse:0.328236 
#> [37]	train-rmse:0.306226 
#> [38]	train-rmse:0.281044 
#> [39]	train-rmse:0.270485 
#> [40]	train-rmse:0.263219 
#> [41]	train-rmse:0.221316 
#> [42]	train-rmse:0.203244 
#> [43]	train-rmse:0.197485 
#> [44]	train-rmse:0.180673 
#> [45]	train-rmse:0.169092 
#> [46]	train-rmse:0.151325 
#> [47]	train-rmse:0.138258 
#> [48]	train-rmse:0.124188 
#> [49]	train-rmse:0.111664 
#> [50]	train-rmse:0.104917 

# Hyperparameter
hyperparam <- c(0.1, 0.2, 1)
n_neighbor <- 10
expand <- 1
block_size <- 10000

# Exposure level
wi <- 0.4

# Estimate GPS for the exposure level
GPS_w = dnorm(wi,
              mean = GPS_m$e_gps_pred,
              sd = GPS_m$e_gps_std, log = TRUE)

# Order data for easy selection
coord_obs = cbind(data$treat, GPS_m$GPS)
y_use <- data$Y

obs_ord <- coord_obs[order(coord_obs[,1]),]
y_use_ord <- y_use[order(coord_obs[,1])]

# compute noise
noise <- estimate_noise_nn(hyperparam = hyperparam,
                           w_obs = data$treat,
                           GPS_obs = GPS_m$GPS,
                           y_obs = y_use_ord,
                           n_neighbor = n_neighbor)

# compute posterior standard deviation
pst_sd <- compute_posterior_sd_nn(hyperparam = hyperparam,
                                  w = wi,
                                  GPS_w = GPS_w,
                                  obs_ord = obs_ord,
                                  sigma2 = noise,
                                  n_neighbor = 20,
                                  expand = 1)