Change-point Detection in Full GP

Calculate the posterior mean of the difference between left- and right-derivatives at an exposure level for the detection of change points.

Usage

compute_rl_deriv_gp(
  w,
  w_obs,
  y_obs,
  GPS_m,
  hyperparam,
  kernel_fn = function(x) exp(-x),
  kernel_deriv_fn = function(x) -exp(-x)
)

Arguments

w

A scalar of exposure level of interest.

w_obs

A vector of observed exposure levels of all samples.

y_obs

A vector of observed outcome values of all samples.

GPS_m

A data.table of GPS vectors.

Column 1: GPS

Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred).

Column 3: Standard deviation of e_gps (e_gps_std)

hyperparam

A vector of hyper-parameters in the GP model.

kernel_fn

The covariance function.

kernel_deriv_fn

The partial derivative of the covariance function.

Value

A numeric value of the posterior mean of the difference between two one-sided derivatives.

Examples

set.seed(847)
data <- generate_synthetic_data(sample_size = 200)
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:8.023071 
#> [2]	train-rmse:6.258433 
#> [3]	train-rmse:4.932584 
#> [4]	train-rmse:3.967626 
#> [5]	train-rmse:3.301938 
#> [6]	train-rmse:2.821067 
#> [7]	train-rmse:2.396043 
#> [8]	train-rmse:2.071846 
#> [9]	train-rmse:1.894523 
#> [10]	train-rmse:1.615702 
#> [11]	train-rmse:1.505742 
#> [12]	train-rmse:1.366276 
#> [13]	train-rmse:1.287354 
#> [14]	train-rmse:1.184969 
#> [15]	train-rmse:1.122623 
#> [16]	train-rmse:1.016618 
#> [17]	train-rmse:0.890574 
#> [18]	train-rmse:0.833882 
#> [19]	train-rmse:0.794390 
#> [20]	train-rmse:0.762039 
#> [21]	train-rmse:0.670639 
#> [22]	train-rmse:0.612766 
#> [23]	train-rmse:0.586782 
#> [24]	train-rmse:0.571285 
#> [25]	train-rmse:0.543004 
#> [26]	train-rmse:0.527778 
#> [27]	train-rmse:0.474710 
#> [28]	train-rmse:0.460427 
#> [29]	train-rmse:0.439429 
#> [30]	train-rmse:0.394121 
#> [31]	train-rmse:0.380530 
#> [32]	train-rmse:0.366848 
#> [33]	train-rmse:0.321964 
#> [34]	train-rmse:0.305011 
#> [35]	train-rmse:0.296717 
#> [36]	train-rmse:0.282010 
#> [37]	train-rmse:0.248859 
#> [38]	train-rmse:0.228316 
#> [39]	train-rmse:0.217543 
#> [40]	train-rmse:0.209351 
#> [41]	train-rmse:0.198866 
#> [42]	train-rmse:0.190064 
#> [43]	train-rmse:0.179161 
#> [44]	train-rmse:0.172966 
#> [45]	train-rmse:0.164995 
#> [46]	train-rmse:0.154928 
#> [47]	train-rmse:0.147185 
#> [48]	train-rmse:0.142516 
#> [49]	train-rmse:0.134590 
#> [50]	train-rmse:0.127147 

wi <- 8.6

val <- compute_rl_deriv_gp(w = wi,
                           w_obs = data$treat,
                           y_obs = data$Y,
                           GPS_m = GPS_m,
                           hyperparam = c(1,1,2))