Compute Posterior Credible Interval

Computes posterior credible interval for requested exposure level

Usage

compute_sd_gp(
  w,
  scaled_obs,
  hyperparam,
  sigma,
  GPS_m,
  kernel_fn = function(x) exp(-x^2)
)

Arguments

w

A scalar of exposure level of interest.

scaled_obs

A matrix of two columns.

First column is the scaled GPS value of all samples (GPS * 1/sqrt(alpha))

Second column is the scaled exposure value of all samples (w * 1/sqrt(beta))

hyperparam

A vector of hyper-parameters for the GP.

First element: alpha

Second element: beta

Third element: gamma/sigma

sigma

A scaler that represents noise.

GPS_m

A data.table of GPS vectors.

Column 1: A vector of estimated GPS evaluated at the observed exposure levels.

Column 2: Estimated conditional means of the exposure given covariates for all samples (e_gps_pred).

Column 3: Estimated conditional standard deviation of the exposure given covariates for all samples (e_gps_std).

kernel_fn

The covariance function of GP.

Value

Posterior credible interval (scaler) for the requested exposure level (w).

Examples

set.seed(284)
#Generate synthetic data
data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)
w_obs <- obs_exposure <- data$treat

# Choose an exposure level to compute CERF
w = 1.2

# Define kernel function
kernel_fn <- function(x) exp(-x^2)

# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:8.126090 
#> [2]	train-rmse:6.346745 
#> [3]	train-rmse:5.160632 
#> [4]	train-rmse:4.237126 
#> [5]	train-rmse:3.639487 
#> [6]	train-rmse:3.173138 
#> [7]	train-rmse:2.860692 
#> [8]	train-rmse:2.588496 
#> [9]	train-rmse:2.294354 
#> [10]	train-rmse:2.076848 
#> [11]	train-rmse:1.911738 
#> [12]	train-rmse:1.698649 
#> [13]	train-rmse:1.627178 
#> [14]	train-rmse:1.490352 
#> [15]	train-rmse:1.341344 
#> [16]	train-rmse:1.225390 
#> [17]	train-rmse:1.130175 
#> [18]	train-rmse:1.009345 
#> [19]	train-rmse:0.955622 
#> [20]	train-rmse:0.892145 
#> [21]	train-rmse:0.820646 
#> [22]	train-rmse:0.780097 
#> [23]	train-rmse:0.762133 
#> [24]	train-rmse:0.717963 
#> [25]	train-rmse:0.659093 
#> [26]	train-rmse:0.589844 
#> [27]	train-rmse:0.579641 
#> [28]	train-rmse:0.523528 
#> [29]	train-rmse:0.491045 
#> [30]	train-rmse:0.466017 
#> [31]	train-rmse:0.403185 
#> [32]	train-rmse:0.350622 
#> [33]	train-rmse:0.312244 
#> [34]	train-rmse:0.282975 
#> [35]	train-rmse:0.247362 
#> [36]	train-rmse:0.223535 
#> [37]	train-rmse:0.205091 
#> [38]	train-rmse:0.197085 
#> [39]	train-rmse:0.183195 
#> [40]	train-rmse:0.166145 
#> [41]	train-rmse:0.160136 
#> [42]	train-rmse:0.151294 
#> [43]	train-rmse:0.136097 
#> [44]	train-rmse:0.128665 
#> [45]	train-rmse:0.116861 
#> [46]	train-rmse:0.113192 
#> [47]	train-rmse:0.107310 
#> [48]	train-rmse:0.101266 
#> [49]	train-rmse:0.096414 
#> [50]	train-rmse:0.085752 

GPS <- GPS_m$GPS

# set hyperparameters
hyperparam <- c(0.1, 0.4, 1)
alpha <- hyperparam[[1]]
beta <- hyperparam[[2]]
g_sigma <- hyperparam[[3]]

# Compute scaled observation data and inverse of covariate matrix.
scaled_obs <- cbind(obs_exposure*sqrt(1/alpha), GPS*sqrt(1/beta))

tentative_sigma <- 0.1

post_sd <- GPCERF:::compute_sd_gp(w = w,
                                 scaled_obs = scaled_obs,
                                 hyperparam = hyperparam,
                                 sigma = tentative_sigma,
                                 GPS_m = GPS_m,
                                 kernel_fn = kernel_fn)