Estimate the Conditional Exposure Response Function using Gaussian Process

Estimates the conditional exposure response function (cerf) using Gaussian Process (gp). The function tune the best match (the lowest covariate balance) for the provided set of hyperparameters.

Usage

estimate_cerf_gp(data, w, GPS_m, params, kernel_fn = function(x) exp(-x^2))

Arguments

data

A data.table of observation data.

Column 1: Outcome (Y)

Column 2: Exposure or treatment (w)

Column 3~m: Confounders (C)

w

A vector of exposure level to compute CERF.

GPS_m

A data.table of GPS vectors.

Column 1: GPS

Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred).

Column 3: Standard deviation of e_gps (e_gps_std)

params

A list of parameters that is required to run the process. These parameters include:

alpha: A scaling factor for the GPS value.

beta: A scaling factor for the exposure value.

g_sigma: A scaling factor for kernel function (gamma/sigma).

tune_app: A tuning approach. Available approaches:

all: try all combinations of hyperparameters. alpha, beta, and g_sigma can be a vector of parameters.

kernel_fn

A kernel function. A default value is a Gaussian Kernel.

Value

A cerf_gp object that includes the following values:

TBD

Examples

set.seed(129)
sim.data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)

# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(sim.data[,-(1:2)]),
                   w.all = as.matrix(sim.data$treat))
#> [1]	train-rmse:8.013595 
#> [2]	train-rmse:6.290675 
#> [3]	train-rmse:5.069039 
#> [4]	train-rmse:4.109136 
#> [5]	train-rmse:3.403551 
#> [6]	train-rmse:2.911113 
#> [7]	train-rmse:2.516932 
#> [8]	train-rmse:2.276880 
#> [9]	train-rmse:2.123440 
#> [10]	train-rmse:2.042455 
#> [11]	train-rmse:1.822187 
#> [12]	train-rmse:1.678244 
#> [13]	train-rmse:1.593165 
#> [14]	train-rmse:1.470786 
#> [15]	train-rmse:1.398698 
#> [16]	train-rmse:1.346891 
#> [17]	train-rmse:1.313656 
#> [18]	train-rmse:1.273869 
#> [19]	train-rmse:1.171202 
#> [20]	train-rmse:1.082402 
#> [21]	train-rmse:0.949546 
#> [22]	train-rmse:0.874509 
#> [23]	train-rmse:0.810788 
#> [24]	train-rmse:0.767233 
#> [25]	train-rmse:0.697216 
#> [26]	train-rmse:0.670815 
#> [27]	train-rmse:0.600684 
#> [28]	train-rmse:0.553597 
#> [29]	train-rmse:0.528699 
#> [30]	train-rmse:0.502762 
#> [31]	train-rmse:0.476917 
#> [32]	train-rmse:0.465998 
#> [33]	train-rmse:0.442204 
#> [34]	train-rmse:0.419775 
#> [35]	train-rmse:0.395772 
#> [36]	train-rmse:0.386138 
#> [37]	train-rmse:0.353329 
#> [38]	train-rmse:0.332995 
#> [39]	train-rmse:0.317206 
#> [40]	train-rmse:0.301822 
#> [41]	train-rmse:0.285412 
#> [42]	train-rmse:0.264272 
#> [43]	train-rmse:0.245202 
#> [44]	train-rmse:0.230332 
#> [45]	train-rmse:0.209733 
#> [46]	train-rmse:0.200581 
#> [47]	train-rmse:0.180731 
#> [48]	train-rmse:0.160498 
#> [49]	train-rmse:0.148313 
#> [50]	train-rmse:0.140261 

# exposure values
w.all = seq(0,20,0.1)

data.table::setDT(sim.data)

cerf_gp_obj <- estimate_cerf_gp(sim.data,
                                w.all,
                                GPS_m,
                                params = list(alpha = c(0.1,0.2,0.4),
                                              beta=0.2,
                                              g_sigma = 1,
                                              tune_app = "all"))
#> Var1 Var2 Var3 
#>  0.1  0.2  1.0 
#> Var1 Var2 Var3 
#>  0.2  0.2  1.0 
#> Var1 Var2 Var3 
#>  0.4  0.2  1.0