Estimate the Conditional Exposure Response Function using Nearest Neighbor Gaussian Process

Estimates the conditional exposure response function (cerf) using the nearest neighbor (nn) Gaussian Process (gp). The function tune the best match (the lowest covariate balance) for the provided set of hyperparameters.

Usage

estimate_cerf_nngp(data, w, GPS_m, params, kernel_fn)

Arguments

data

A data.table of observation data.

Column 1: Outcome (Y)

Column 2: Exposure or treatment (w)

Column 3~m: Confounders (C)

w

A vector of exposure level to compute CERF.

GPS_m

A data.table of GPS vectors.

Column 1: GPS

Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred).

Column 3: Standard deviation of e_gps (e_gps_std)

params

A list of parameters that is required to run the process. These parameters include:

alpha: A scaling factor for the GPS value.

beta: A scaling factor for the exposure value.

g_sigma: A scaling factor for kernel function (gamma/sigma).

tune_app: A tuning approach. Available approaches:

all: try all combinations of hyperparameters.

expand: Scaling factor to determine the total number of nearest neighbors. The total is 2*expand*n.neighbour.

n_neighbor: Number of nearest neighbors on one side.

block_size: Number of samples included in a computation block. Mainly used to balance the speed and memory requirement. Larger block_size is faster, but requires more memory. alpha, beta, and g_sigma can be a vector of parameters.

kernel_fn

A kernel function. A default value is a Gaussian Kernel.

Value

#' A cerf_nngp object that includes the following values:

TBD

Examples

set.seed(19)
sim.data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)
# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(sim.data[,-(1:2)]),
                   w.all = as.matrix(sim.data$treat))
#> [1]	train-rmse:7.836669 
#> [2]	train-rmse:6.134167 
#> [3]	train-rmse:4.915488 
#> [4]	train-rmse:3.988384 
#> [5]	train-rmse:3.336050 
#> [6]	train-rmse:2.859457 
#> [7]	train-rmse:2.479927 
#> [8]	train-rmse:2.112485 
#> [9]	train-rmse:1.843607 
#> [10]	train-rmse:1.648466 
#> [11]	train-rmse:1.506709 
#> [12]	train-rmse:1.381182 
#> [13]	train-rmse:1.257668 
#> [14]	train-rmse:1.156825 
#> [15]	train-rmse:1.116339 
#> [16]	train-rmse:1.041766 
#> [17]	train-rmse:0.910545 
#> [18]	train-rmse:0.880408 
#> [19]	train-rmse:0.809074 
#> [20]	train-rmse:0.728130 
#> [21]	train-rmse:0.638991 
#> [22]	train-rmse:0.601759 
#> [23]	train-rmse:0.550966 
#> [24]	train-rmse:0.531337 
#> [25]	train-rmse:0.482819 
#> [26]	train-rmse:0.452625 
#> [27]	train-rmse:0.431795 
#> [28]	train-rmse:0.402634 
#> [29]	train-rmse:0.387928 
#> [30]	train-rmse:0.367107 
#> [31]	train-rmse:0.346419 
#> [32]	train-rmse:0.335399 
#> [33]	train-rmse:0.326148 
#> [34]	train-rmse:0.299594 
#> [35]	train-rmse:0.282238 
#> [36]	train-rmse:0.264984 
#> [37]	train-rmse:0.249363 
#> [38]	train-rmse:0.244614 
#> [39]	train-rmse:0.228311 
#> [40]	train-rmse:0.218213 
#> [41]	train-rmse:0.207740 
#> [42]	train-rmse:0.194086 
#> [43]	train-rmse:0.189707 
#> [44]	train-rmse:0.172412 
#> [45]	train-rmse:0.159344 
#> [46]	train-rmse:0.143543 
#> [47]	train-rmse:0.133508 
#> [48]	train-rmse:0.124262 
#> [49]	train-rmse:0.112816 
#> [50]	train-rmse:0.109303 
# exposure values
w.all <- seq(0,20,0.5)
data.table::setDT(sim.data)
cerf_nngp_obj <- estimate_cerf_nngp(sim.data,
                                    w.all,
                                    GPS_m,
                                    params = list(alpha = c(0.1,0.2),
                                                  beta = 0.2,
                                                  g_sigma = 1,
                                                  tune_app = "all",
                                                  n_neighbor = 20,
                                                  expand = 1,
                                                  block_size = 1e4))