Estimate the CERF with the nnGP Model

Estimates the posterior mean of the conditional exposure response function at specified exposure levels with nnGP.

Usage

estimate_mean_sd_nn(
  hyperparam,
  sigma2,
  w_obs,
  w,
  y_obs,
  GPS_m,
  n_neighbor = 50,
  expand = 2,
  block_size = 2000
)

Arguments

hyperparam

A set of hyperparameters for the nnGP.

sigma2

A scaler representing sigma^2.

w_obs

A vector of observed exposure levels.

w

A vector of exposure levels at which the CERF is estimated.

y_obs

A vector of observed outcome values.

GPS_m

A data.table of GPS vectors.

Column 1: GPS

Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred).

Column 3: Standard deviation of e_gps (e_gps_std)

n_neighbor

Number of nearest neighbours on one side (see also expand).

expand

Scaling factor to determine the total number of nearest neighbours. The total is 2*expand*n.neighbour.

block_size

Number of samples included in a computation block. Mainly used to balance the speed and memory requirement. Larger block.size is faster, but requires more memory.

Value

A list of 2 elements, including:

the returned value from compute_posterior_m_nn

the returned value from compute_posterior_sd_nn

Examples

set.seed(86)
data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)

# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:7.975127 
#> [2]	train-rmse:6.283342 
#> [3]	train-rmse:5.081566 
#> [4]	train-rmse:4.248909 
#> [5]	train-rmse:3.548358 
#> [6]	train-rmse:3.048069 
#> [7]	train-rmse:2.699687 
#> [8]	train-rmse:2.327539 
#> [9]	train-rmse:2.158164 
#> [10]	train-rmse:1.955130 
#> [11]	train-rmse:1.709535 
#> [12]	train-rmse:1.601056 
#> [13]	train-rmse:1.450402 
#> [14]	train-rmse:1.367742 
#> [15]	train-rmse:1.281822 
#> [16]	train-rmse:1.171397 
#> [17]	train-rmse:1.022067 
#> [18]	train-rmse:0.875262 
#> [19]	train-rmse:0.796402 
#> [20]	train-rmse:0.691177 
#> [21]	train-rmse:0.637054 
#> [22]	train-rmse:0.587311 
#> [23]	train-rmse:0.561769 
#> [24]	train-rmse:0.540759 
#> [25]	train-rmse:0.501040 
#> [26]	train-rmse:0.464191 
#> [27]	train-rmse:0.432223 
#> [28]	train-rmse:0.391812 
#> [29]	train-rmse:0.348828 
#> [30]	train-rmse:0.327255 
#> [31]	train-rmse:0.280866 
#> [32]	train-rmse:0.263177 
#> [33]	train-rmse:0.244668 
#> [34]	train-rmse:0.233667 
#> [35]	train-rmse:0.219347 
#> [36]	train-rmse:0.208436 
#> [37]	train-rmse:0.197501 
#> [38]	train-rmse:0.188959 
#> [39]	train-rmse:0.173470 
#> [40]	train-rmse:0.168085 
#> [41]	train-rmse:0.151730 
#> [42]	train-rmse:0.140356 
#> [43]	train-rmse:0.127010 
#> [44]	train-rmse:0.119286 
#> [45]	train-rmse:0.106564 
#> [46]	train-rmse:0.100877 
#> [47]	train-rmse:0.096880 
#> [48]	train-rmse:0.092400 
#> [49]	train-rmse:0.084644 
#> [50]	train-rmse:0.080494 

# Hyperparameter
hyperparam <- c(0.1, 0.2, 1)
n_neighbor <- 15
expand <- 1
block_size <- 10000

# compute noise
noise <- estimate_noise_nn(hyperparam = hyperparam,
                           w_obs = data$treat,
                           GPS_obs = GPS_m$GPS,
                           y_obs = data$Y,
                           n_neighbor = n_neighbor)

# compute posterior mean and standard deviation for vector of w.
w <- seq(0,20,1)
val <- estimate_mean_sd_nn(hyperparam = hyperparam,
                           sigma2 = noise,
                           w_obs = data$treat,
                           w = w,
                           y_obs = data$Y,
                           GPS_m = GPS_m,
                           n_neighbor = n_neighbor,
                           expand = expand,
                           block_size = block_size)