Estimate the Standard Deviation of the Nugget Term in Full Gaussian Process

Estimates the standard deviations of the nugget term in full GP by calculating the standard deviations of the residuals.

Usage

estimate_noise(hyperparam, data, GPS)

Arguments

hyperparam

A vector of hyper-parameter values for the full GP.

data

A data.table of observation data.

Column 1: Outcome (Y)

Column 2: Exposure or treatment (w)

Column 3~m: Confounders (C)

GPS

A vector of estimated GPS at the observed exposure levels.

Value

A scalar of estimated standard deviation of the nugget term in full GP.

Examples

set.seed(109)
data <- generate_synthetic_data(sample_size = 100, gps_spec = 3)
data.table::setDT(data)

# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:8.557412 
#> [2]	train-rmse:6.544529 
#> [3]	train-rmse:5.044576 
#> [4]	train-rmse:3.991472 
#> [5]	train-rmse:3.228516 
#> [6]	train-rmse:2.569900 
#> [7]	train-rmse:2.094692 
#> [8]	train-rmse:1.731734 
#> [9]	train-rmse:1.470229 
#> [10]	train-rmse:1.274962 
#> [11]	train-rmse:1.054901 
#> [12]	train-rmse:0.888915 
#> [13]	train-rmse:0.753735 
#> [14]	train-rmse:0.648757 
#> [15]	train-rmse:0.572981 
#> [16]	train-rmse:0.502921 
#> [17]	train-rmse:0.443061 
#> [18]	train-rmse:0.395692 
#> [19]	train-rmse:0.347662 
#> [20]	train-rmse:0.305686 
#> [21]	train-rmse:0.273674 
#> [22]	train-rmse:0.242320 
#> [23]	train-rmse:0.221018 
#> [24]	train-rmse:0.191208 
#> [25]	train-rmse:0.176630 
#> [26]	train-rmse:0.163489 
#> [27]	train-rmse:0.142951 
#> [28]	train-rmse:0.122548 
#> [29]	train-rmse:0.112130 
#> [30]	train-rmse:0.096368 
#> [31]	train-rmse:0.084064 
#> [32]	train-rmse:0.070788 
#> [33]	train-rmse:0.062368 
#> [34]	train-rmse:0.057281 
#> [35]	train-rmse:0.051983 
#> [36]	train-rmse:0.045176 
#> [37]	train-rmse:0.039839 
#> [38]	train-rmse:0.035014 
#> [39]	train-rmse:0.029088 
#> [40]	train-rmse:0.024278 
#> [41]	train-rmse:0.020719 
#> [42]	train-rmse:0.018624 
#> [43]	train-rmse:0.016722 
#> [44]	train-rmse:0.015417 
#> [45]	train-rmse:0.013064 
#> [46]	train-rmse:0.011746 
#> [47]	train-rmse:0.010214 
#> [48]	train-rmse:0.008645 
#> [49]	train-rmse:0.007379 
#> [50]	train-rmse:0.006359 

hyperparam <- c(0.1, 0.2, 1)

noise_est <- estimate_noise(hyperparam, data, GPS_m$GPS)