Find the Optimal Hyper-parameter for the Nearest Neighbor Gaussian Process

Computes covariate balance for each combincation of provided hyper-parameters and selects the hyper-parameter values that minimizes the covariate balance.

Usage

find_optimal_nn(
  w_obs,
  w,
  y_obs,
  GPS_m,
  design_mt,
  hyperparams = expand.grid(seq(0.5, 4.5, 1), seq(0.5, 4.5, 1), seq(0.5, 4.5, 1)),
  n_neighbor = 50,
  expand = 2,
  block_size = 2000
)

Arguments

w_obs

A vector of the observed exposure levels.

w

A vector of exposure levels at which CERF will be estimated.

y_obs

A vector of observed outcomes

GPS_m

A data.table of GPS vectors.

Column 1: GPS

Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred).

Column 3: Standard deviation of e_gps (e_gps_std)

design_mt

The covariate matrix of all samples (intercept excluded).

hyperparams

A matrix of candidate values of the hyper-parameters, each row contains a set of values of all hyper-parameters.

n_neighbor

Number of nearest neighbors on one side (see also expand).

expand

Scaling factor to determine the total number of nearest neighbours. The total is 2*expand*n_neighbor.

block_size

Number of samples included in a computation block. Mainly used to balance the speed and memory requirement. Larger block_size is faster, but requires more memory.

Value

Estimated covariate balance scores for the grid of hyper-parameter values considered in hyperparams.

Examples

set.seed(89)
data <- generate_synthetic_data(sample_size = 200, gps_spec = 3)

# Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))
#> [1]	train-rmse:8.206952 
#> [2]	train-rmse:6.394042 
#> [3]	train-rmse:5.031031 
#> [4]	train-rmse:4.074021 
#> [5]	train-rmse:3.325362 
#> [6]	train-rmse:2.738510 
#> [7]	train-rmse:2.303424 
#> [8]	train-rmse:2.037422 
#> [9]	train-rmse:1.778079 
#> [10]	train-rmse:1.627036 
#> [11]	train-rmse:1.513744 
#> [12]	train-rmse:1.383442 
#> [13]	train-rmse:1.312099 
#> [14]	train-rmse:1.191915 
#> [15]	train-rmse:1.096559 
#> [16]	train-rmse:1.035653 
#> [17]	train-rmse:0.961320 
#> [18]	train-rmse:0.936646 
#> [19]	train-rmse:0.902146 
#> [20]	train-rmse:0.851613 
#> [21]	train-rmse:0.814890 
#> [22]	train-rmse:0.776779 
#> [23]	train-rmse:0.741845 
#> [24]	train-rmse:0.715681 
#> [25]	train-rmse:0.655899 
#> [26]	train-rmse:0.583473 
#> [27]	train-rmse:0.530346 
#> [28]	train-rmse:0.467600 
#> [29]	train-rmse:0.433730 
#> [30]	train-rmse:0.412744 
#> [31]	train-rmse:0.381812 
#> [32]	train-rmse:0.340381 
#> [33]	train-rmse:0.333415 
#> [34]	train-rmse:0.315124 
#> [35]	train-rmse:0.292065 
#> [36]	train-rmse:0.269160 
#> [37]	train-rmse:0.245788 
#> [38]	train-rmse:0.226206 
#> [39]	train-rmse:0.207609 
#> [40]	train-rmse:0.196317 
#> [41]	train-rmse:0.179336 
#> [42]	train-rmse:0.167567 
#> [43]	train-rmse:0.156670 
#> [44]	train-rmse:0.149022 
#> [45]	train-rmse:0.138442 
#> [46]	train-rmse:0.132194 
#> [47]	train-rmse:0.128503 
#> [48]	train-rmse:0.112246 
#> [49]	train-rmse:0.106002 
#> [50]	train-rmse:0.100278 

# Hyperparameter
hyperparam <- c(0.1, 0.2, 1)
n_neighbor <- 10
expand <- 1
block_size <- 10000

# compute posterior mean and standard deviation for vector of w.
w <- seq(0,20,2)
design_mt <- model.matrix(~.-1, data = data[, 3:ncol(data)])

hyperparam_grid <- expand.grid(seq(0.5,2.5,1),
                               seq(0.4,0.6,0.2),
                               seq(0.5,1.5,1))

optimal_cb <- find_optimal_nn(w_obs = data$treat,
                              w = w,
                              y_obs = data$Y,
                              GPS_m = GPS_m,
                              design_mt = design_mt,
                              hyperparams = hyperparam_grid,
                              n_neighbor = 50, expand = 2, block_size = 2e3)
#> Var1 Var2 Var3 
#>  0.5  0.4  0.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  1.5  0.4  0.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  2.5  0.4  0.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  0.5  0.6  0.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  1.5  0.6  0.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  2.5  0.6  0.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  0.5  0.4  1.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  1.5  0.4  1.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  2.5  0.4  1.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  0.5  0.6  1.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  1.5  0.6  1.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20
#> Var1 Var2 Var3 
#>  2.5  0.6  1.5 
#> [1] 0
#> [1] 2
#> [1] 4
#> [1] 6
#> [1] 8
#> [1] 10
#> [1] 12
#> [1] 14
#> [1] 16
#> [1] 18
#> [1] 20