Compute mean, confidence interval (?), and covariate balance in Full Gaussian Process (GP)

Calculates the induced covariate balance associated with one hyper-parameter configuration in full GP.

Usage

compute_m_sigma(
  hyperparam,
  data,
  w,
  GPS_m,
  nthread = 1,
  kernel_fn = function(x) exp(-x^2)
)

Arguments

hyperparam

A vector of values of hyper-parameters.

First element: alpha

Second element: beta

Third element: g_sigma (gamma/sigma)

data

A data.table containing all data including outcome, exposure and covariates. In the following order:

Column 1: Outcome (Y)

Column 2: Exposure or treatment (w)

Column 3~m: Confounders (C)

w

A vector of exposure levels at which the CERF is estimated.

GPS_m

A data.table of GPS vectors.

Column 1: A vector of estimated GPS evaluated at the observed exposure levels.

Column 2: Estimated conditional means of the exposure given covariates for all samples (e_gps_pred).

Column 3: Estimated conditional standard deviation of the exposure given covariates for all samples (e_gps_std).

nthread

An integer value that represents the number of threads to be used by internal packages.

kernel_fn

The covariance function of GP.

Value

A list containing two elements: 1) a vector of absolute weighted correlation of each covariate to the exposure, which is the metric for covariate balance and 2) the estimated CERF at w.all based on the hyper-parameter values in param.

Examples

set.seed(912)
data <- generate_synthetic_data(sample_size = 250, gps_spec = 3)

w.all = seq(0,20,0.1)

data.table::setDT(data)

#Estimate GPS function
GPS_m <- train_GPS(cov.mt = as.matrix(data[,-(1:2)]),
                   w.all = as.matrix(data$treat))

tune_res <- compute_m_sigma(hyperparam = c(0.09, 0.09, 10),
                            data = data,
                            w = w.all,
                            GPS_m = GPS_m,
                            nthread = 1)

gp.cerf <- tune_res$est