Compute mean, credible interval, and covariate balance in full Gaussian process (GP)

Calculates the induced covariate balance associated with one hyper-parameter configuration in full GP.

Usage

compute_m_sigma(
  hyperparam,
  data,
  w,
  gps_m,
  tuning,
  kernel_fn = function(x) exp(-x^2)
)

Arguments

hyperparam

A vector of values of hyper-parameters.

First element: alpha

Second element: beta

Third element: g_sigma (gamma / sigma)

data

A data.frame containing all data including outcome, exposure and covariates. In the following order:

Column 1: Outcome (Y)

Column 2: Exposure or treatment (w)

Column 3~m: Confounders (C)

w

A vector of exposure levels at which the CERF is estimated.

gps_m

An S3 gps object including: gps: A data.frame of GPS vectors. - Column 1: GPS - Column 2: Prediction of exposure for covariate of each data sample (e_gps_pred). - Column 3: Standard deviation of e_gps (e_gps_std) used_params: - dnorm_log: TRUE or FLASE

tuning

The function is used for parameter tuning (default = TRUE) or estimation (FALSE)

kernel_fn

The covariance function of GP.

Value

A list containing two elements:

A vector of absolute weighted correlation of each covariate to the exposure, which is the metric for covariate balance

An estimated CERF at w_all based on the hyper-parameter values in param.