Calculate posterior standard deviations for nnGP model

Calculates the posterior standard deviation of a point on the CERF based on the nnGP model.

Usage

compute_posterior_sd_nn(
  hyperparam,
  w,
  GPS_w,
  obs_ord,
  sigma2,
  kernel_fn = function(x) exp(-x^2),
  n_neighbor = 10,
  expand = 1,
  block_size = 10000
)

Arguments

hyperparam

The values of hyperparameters in the GP model.

w

The exposure level for the point of interest on the CERF.

GPS_w

The GPS for all samples when their exposure levels are set at w.

obs_ord

A matrix of two columns. The first column is the observed exposure levels of all samples; the second is the GPS at the observed exposure levels. The rows are in ascending order for the first column.

sigma2

A scaler representing sigma^2.

kernel_fn

The covariance function of the GP.

n_neighbor

Number of nearest neighbors on one side (see also expand).

expand

A scaling factor to determine the total number of nearest neighbors. The total is 2*expand*n_neighbor.

block_size

Number of samples included in a computation block. Mainly used to balance the speed and memory requirement. Larger block_size is faster, but requires more memory.

Value

The posterior standard deviation of the estimated CERF at w.